+++ src/RcppExports.cpp
@@ -26,13 +26,13 @@ BEGIN_RCPP
 END_RCPP
 }
 // readSparseData
-List readSparseData(std::string filename, size_t skipBytes, bool verbose, bool zeroBased);
+List readSparseData(std::string filename, unsigned long skipBytes, bool verbose, bool zeroBased);
 RcppExport SEXP SVMBridge_readSparseData(SEXP filenameSEXP, SEXP skipBytesSEXP, SEXP verboseSEXP, SEXP zeroBasedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
     Rcpp::traits::input_parameter< std::string >::type filename(filenameSEXP);
-    Rcpp::traits::input_parameter< size_t >::type skipBytes(skipBytesSEXP);
+    Rcpp::traits::input_parameter< unsigned long >::type skipBytes(skipBytesSEXP);
     Rcpp::traits::input_parameter< bool >::type verbose(verboseSEXP);
     Rcpp::traits::input_parameter< bool >::type zeroBased(zeroBasedSEXP);
     __result = Rcpp::wrap(readSparseData(filename, skipBytes, verbose, zeroBased));
@@ -40,7 +40,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // writeSparseData
-List writeSparseData(std::string filename, NumericMatrix X, NumericMatrix Y, size_t skipBytes, bool verbose, bool zeroBased);
+List writeSparseData(std::string filename, NumericMatrix X, NumericMatrix Y, unsigned long skipBytes, bool verbose, bool zeroBased);
 RcppExport SEXP SVMBridge_writeSparseData(SEXP filenameSEXP, SEXP XSEXP, SEXP YSEXP, SEXP skipBytesSEXP, SEXP verboseSEXP, SEXP zeroBasedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
@@ -48,7 +48,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< std::string >::type filename(filenameSEXP);
     Rcpp::traits::input_parameter< NumericMatrix >::type X(XSEXP);
     Rcpp::traits::input_parameter< NumericMatrix >::type Y(YSEXP);
-    Rcpp::traits::input_parameter< size_t >::type skipBytes(skipBytesSEXP);
+    Rcpp::traits::input_parameter< unsigned long >::type skipBytes(skipBytesSEXP);
     Rcpp::traits::input_parameter< bool >::type verbose(verboseSEXP);
     Rcpp::traits::input_parameter< bool >::type zeroBased(zeroBasedSEXP);
     __result = Rcpp::wrap(writeSparseData(filename, X, Y, skipBytes, verbose, zeroBased));
